Refactor company attribute and workspace outputs
Refactored how the 'company' attribute in 'user.rb' is built by removing the .data method. In 'workspace.rb', the way outputs are handled has been updated to make the output the attribute's name, resulting in a more dynamic approach. A line has also been added in 'user_spec.rb' to inspect the changes made in 'company'.