Implement output formatter option for attributes
This commit introduces an `output` option for attribute definitions in the user model which represents an output formatter. It can accept a Proc and when provided, will call this Proc with formatted value as input for better control over the output formatting of the attribute value. This adds flexibility in customizing the serialization of attribute values.