Replace option keyword 'as' with 'to' in attribute names
The keyword used to rename attributes was changed from 'as' to 'to' for better readability and understanding. This modification has been applied in user_related and attributes-related files, affecting attributes related to users' personal details, address and geolocation data. It also required corresponding changes on handle option fetching in data factory and serialization.