Refactor attribute type parsing in service factory
This update enhances the attribute parsing logic in service factory. If the attribute has an 'as' option defined, its value is now transformed to the specified type, which can be Date, Time, DateTime, String, Integer, or Float. The logic was also simplified for handling 'consists_of' type in input_internal_name.