Add HashToObject class and integrate into serialization process
Introduced a new HashToObject class in the Datory::Attributes::Serialization module. This class turns hash into object which can be used in serialization process. Updated the serialize method in the Callable module to convert Hash and Array of Hashes to objects using the new HashToObject class. Refactored the user_spec test file to test the serialize method with objects and hash.