Add unused attribute to user spec test
An unused attribute has been added to the user spec test in `example1/user_spec.rb`. This attribute, labelled as "unused_attribute", includes a note remarking its redundancy and its purpose of testing success and ignore. This addition is to create awareness about the presence of this attribute and its contribution to the tests.