Refactor serialization methods and disable rubocop rule
The serialization methods in HashToObject and Callable modules were refactored. The HashToObject 'build' method was replaced with a more comprehensive 'prepare' method. The Callable module's 'serialize' method was simplified by utilizing the new 'prepare' method. Additionally, the RSpec/DescribedClass rubocop rule was disabled in a user_spec file for the User class.