Refactor attribute type extraction in serializator
Switched from fetching the ':type' option to fetching the ':from' option when retrieving the attribute's class from the options hash in the serialization process. This refactoring is required as the ':from' key more accurately represents the type of data we wish to serialize, particularly when dealing with sets and arrays.