Replace `build` methods with `deserialize` in datory context and attributes
The commit includes renaming and replacement of `build` methods with `deserialize` across various files. This is part of a larger refactor which replaces usage of `build` methods as the main method of object creation. These changes simplify the object creation and deserialization process. It also includes the removal of build and build! methods and associated tests involving these in the `user_spec` file.