Add serialization service builder and factory
The new classes, ServiceBuilder and ServiceFactory, have been introduced in the Datory::Attributes::Serialization module. ServiceBuilder creates a service instance to serialize particular attributes of a model. ServiceFactory is responsible for creating a model specific service class for serialization if not already present. The unnecessary attributes are also removed during the process.