Add Form class and update Callable and Product spec
A new Form class is introduced in Datory::Attributes, containing methods related to data serialization and validation. Callable module is updated to include a form method, which instantiates a new Form object. Product spec is also updated to add test cases for this functionality, checking validity and serialization of data.