Add update method to form model and related tests
Updated the form model by adding an `update` method that allows attributes to be altered. Exceptions are raised when this method is used with a collection. Accordingly, tests have been added to the user and product specs to ensure correct functionality and error handling. The 'user' references in the product_spec.rb have also been corrected to 'product'.