Update Datory methods to use bang notation
The methods 'uuid', 'money', 'duration', 'time' and 'datetime' have been updated to use bang notation, indicating the operation will mutate their caller. The non-bang methods ('uuid', 'money', 'duration', 'time', and 'datetime') have been aliased to the bang methods and are expected to be deprecated in the future. The changes have been propagated to the Example1 and Example2 modules.