Add optional duration attribute to product model
An optional `duration` attribute has been added to the product model, allowing for a possible installment duration attribute. This attribute allows for either a String or NilClass, and it can be converted into an ActiveSupport::Duration or NilClass. Tests and examples have been modified accordingly to reflect this new change.